refactor move makeSurveyPost to lib
