+++ src/bot.js
@@ -1,5 +1,5 @@
-import surveyToText from "./lib/surveyToText.js";
 import { createSurvey } from "./lib/createSurvey.js";
+import { makeSurveyPost } from "./lib/makeSurveyPost.js";
 import {
   maxResponsesForMultiResponsePerUser,
   create_multi_cmd,
@@ -462,36 +462,6 @@ const startAutoPosting = async (client, redisClient) => {
         }
       }
     }
-  }
-};
-
-const makeSurveyPost = async (redisClient, surveyName) => {
-  if (!(await redisClient.sIsMember("surveys", surveyName))) {
-    return [{ content: "There is no survey with that name", ephemeral: true }];
-  } else {
-    const [msg, files] = await surveyToText(redisClient, surveyName);
-
-    const lines = msg.split("\n");
-    const chunks = [];
-    let chunk = "";
-    for (const line of lines) {
-      const chunkWithLine = chunk + "\n" + line;
-      if (chunkWithLine.length > 2000) {
-        chunks.push(chunk);
-        chunk = "";
-      }
-      chunk = chunk + "\n" + line;
-    }
-    chunks.push(chunk);
-
-    return chunks.map((chunk, i) => {
-      console.log("making chunk", i, chunk.length);
-      const toSend = { content: chunk };
-      if (i == chunks.length - 1) {
-        toSend.files = files;
-      }
-      return toSend;
-    });
   }
 };
